+++ src/main/java/org/intellij/sdk/project/model/DebuggerWindow.java
@@ -5,19 +5,19 @@ package org.intellij.sdk.project.model;
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
 import javax.swing.tree.DefaultTreeModel;
-import org.intellij.sdk.project.model.components.ButtonHandler;
+import org.intellij.sdk.project.model.components.handlers.ButtonHandler;
 import org.intellij.sdk.project.model.components.ButtonType;
-import org.intellij.sdk.project.model.components.ClearHandler;
-import org.intellij.sdk.project.model.components.CollapseTreeHandler;
-import org.intellij.sdk.project.model.components.DeleteHandler;
-import org.intellij.sdk.project.model.components.DiffHandler;
-import org.intellij.sdk.project.model.components.DiffSavedNodesView;
-import org.intellij.sdk.project.model.components.ExpandTreeHandler;
-import org.intellij.sdk.project.model.components.SaveHandler;
-import org.intellij.sdk.project.model.components.SavedNodesSettingsView;
-import org.intellij.sdk.project.model.components.SnapHandler;
-import org.intellij.sdk.project.model.components.ToolHandler;
-import org.intellij.sdk.project.model.components.UploadNodesView;
+import org.intellij.sdk.project.model.components.handlers.ClearHandler;
+import org.intellij.sdk.project.model.components.handlers.CollapseTreeHandler;
+import org.intellij.sdk.project.model.components.handlers.DeleteHandler;
+import org.intellij.sdk.project.model.components.handlers.DiffHandler;
+import org.intellij.sdk.project.model.components.views.DiffSavedNodesView;
+import org.intellij.sdk.project.model.components.handlers.ExpandTreeHandler;
+import org.intellij.sdk.project.model.components.handlers.SaveHandler;
+import org.intellij.sdk.project.model.components.views.SavedNodesSettingsView;
+import org.intellij.sdk.project.model.components.handlers.SnapHandler;
+import org.intellij.sdk.project.model.components.handlers.ToolHandler;
+import org.intellij.sdk.project.model.components.views.UploadNodesView;
 import org.intellij.sdk.project.model.listeners.DebuggerTreeModelListener;
 import org.jetbrains.annotations.NotNull;
 import com.intellij.openapi.project.Project;
