+++ src/main/java/org/intellij/sdk/project/model/components/handlers/ClearHandler.java
@@ -1,4 +1,4 @@
-package org.intellij.sdk.project.model.components;
+package org.intellij.sdk.project.model.components.handlers;
 
 import static org.intellij.sdk.project.model.constants.TextConstants.CLEARED_SNAP_MESSAGE;
 import static org.intellij.sdk.project.model.constants.TextConstants.WINDOW_IS_ALREADY_CLEAR;
