+++ src/main/java/org/intellij/sdk/project/model/components/handlers/CollapseTreeHandler.java
@@ -1,4 +1,4 @@
-package org.intellij.sdk.project.model.components;
+package org.intellij.sdk.project.model.components.handlers;
 
 import javax.swing.*;
 import javax.swing.tree.DefaultTreeModel;
