+++ src/main/java/org/intellij/sdk/project/model/components/views/SavedNodesSettingsView.java
@@ -1,4 +1,4 @@
-package org.intellij.sdk.project.model.components;
+package org.intellij.sdk.project.model.components.views;
 
 import static org.intellij.sdk.project.model.services.ParserService.convertNodeToString;
 
@@ -7,9 +7,14 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import javax.swing.*;
+import javax.swing.border.EmptyBorder;
 import javax.swing.tree.DefaultTreeModel;
+import org.intellij.sdk.project.model.components.handlers.DeleteHandler;
+import org.intellij.sdk.project.model.components.handlers.SaveHandler;
+import org.intellij.sdk.project.model.constants.MessageDialogues;
 import org.intellij.sdk.project.model.services.PersistencyService;
 import org.intellij.sdk.project.model.xnodes.DebugNode;
 import org.jetbrains.annotations.NotNull;
@@ -20,28 +25,22 @@ import com.intellij.openapi.ui.DialogWrapper;
 import com.intellij.ui.components.JBList;
 import com.intellij.ui.components.JBScrollPane;
 
-public class UploadNodesView extends DialogWrapper {
+public class SavedNodesSettingsView extends DialogWrapper {
     private static final PersistencyService persistencyService = ServiceManager.getService(PersistencyService.class);
     private Project project;
+    private SaveHandler saveHandler;
+    private DeleteHandler deleteHandler;
     private DefaultTreeModel treeModel;
     private JBList<String> keysList;
 
-    public UploadNodesView(@NotNull Project project, DefaultTreeModel treeModel) {
+    public SavedNodesSettingsView(@NotNull Project project, SaveHandler saveHandler, DeleteHandler deleteHandler, DefaultTreeModel treeModel) {
         super(true); // use current window as parent
         this.project = project;
+        this.saveHandler = saveHandler;
+        this.deleteHandler = deleteHandler;
         this.treeModel = treeModel;
         setTitle("Manage Saved Nodes");
         init();
-    }
-
-    @Override
-    protected void doOKAction() {
-        super.doOKAction();
-        if (!this.keysList.isSelectionEmpty()) {
-            String selectedValue = this.keysList.getSelectedValue();
-            DebugNode debugNode = persistencyService.getNodes().get(selectedValue);
-            this.treeModel.setRoot(debugNode);
-        }
     }
 
     @Nullable
@@ -64,14 +63,68 @@ public class UploadNodesView extends DialogWrapper {
         dialogPanel.add(scrollableKeysPanel, gbc);
 
         gbc.gridx = 1;
+        gbc.weightx = 0;
+        dialogPanel.add(getButtonsPanel(), gbc);
+
+        gbc.gridx = 2;
         gbc.weightx = 1;
         dialogPanel.add(scrollableNodesPanel, gbc);
         return dialogPanel;
     }
 
+    private JPanel getButtonsPanel() {
+        JPanel panel = new JPanel();
+        panel.setBorder(new EmptyBorder(50, 5, 0, 5));
+        BoxLayout boxLayout = new BoxLayout(panel, BoxLayout.Y_AXIS);
+        panel.setLayout(boxLayout);
+        JButton deleteButton = new JButton("Delete");
+        deleteButton.setEnabled(false);
+
+        deleteButton.addActionListener(e -> deleteNodeAndRefreshJList(this.keysList.getSelectedValue(), false));
+
+        panel.add(deleteButton);
+        JButton renameButton = new JButton("Rename");
+        renameButton.setEnabled(false);
+        this.keysList.addListSelectionListener(e -> {
+            if (e.getFirstIndex() > -1) {
+                deleteButton.setEnabled(true);
+                renameButton.setEnabled(true);
+            } else {
+                deleteButton.setEnabled(false);
+                renameButton.setEnabled(false);
+            }
+        });
+
+        renameButton.addActionListener(e -> {
+            String selectedKey = this.keysList.getSelectedValue();
+            Map<String, DebugNode> nodes = persistencyService.getNodes();
+            DebugNode debugNode = nodes.get(selectedKey);
+            String newNodeName = MessageDialogues.getRenameDialogue(this.project,null);
+            while (nodes.containsKey(newNodeName)) {
+                newNodeName = MessageDialogues.getRenameDialogue(this.project, String.format("\"%s\" already exists", newNodeName));
+            }
+            if (Objects.nonNull(newNodeName)) {
+                this.saveHandler.savedNode(newNodeName, debugNode);
+                deleteNodeAndRefreshJList(selectedKey, true);
+            }
+        });
+
+        panel.add(renameButton);
+        return panel;
+    }
+
+    private void deleteNodeAndRefreshJList(String selectedNode, boolean withOutDialogue) {
+        deleteHandler.handle(treeModel, selectedNode, withOutDialogue);
+        DefaultListModel<String> model = new DefaultListModel<>();
+        model.addAll(persistencyService.getNodes().keySet().stream().sorted().collect(Collectors.toList()));
+        this.keysList.setModel(model);
+    }
+
     @NotNull
     private JScrollPane getScrollableKeysPanel(Map<String, DebugNode> nodes, JScrollPane scrollableNodesPanel) {
         JScrollPane scrollableKeysPanel = new JBScrollPane();
+//        scrollableKeysPanel.setPreferredSize(new Dimension(350, 1100));
+//        scrollableKeysPanel.setSize(new Dimension(350, 1100));
         scrollableKeysPanel.setMinimumSize(new Dimension(350, 1100));
         scrollableKeysPanel.setBorder(BorderFactory.createTitledBorder("Saved Node names"));
